updated fill forwards
